refactor(activemodel,activerecord): split TS-only attribute tests, converge indexParts order, check-constraint defined_for?, FK change-column migration
- activemodel: move the 22 TS-only `it(...)` out of the mirrored
  `attribute.test.ts` into `attribute.trails.test.ts` (RFC 0115).
  attribute_test.rb: OK 33, Extra 22 -> 0; matched total 963 unchanged.
- schema-dumper: `indexParts` appends `include` before `nullsNotDistinct`,
  matching schema_dumper.rb:265-280.
- CheckConstraintDefinition#isDefinedFor ports the `**options` residual arm
  (`options.slice(*self.options.keys)` + `to_s` compare),
  schema_definitions.rb:189-195.
- migration/foreign-key: ForeignKeyChangeColumnTest creates/drops its tables
  through a CreateRocketsMigration port (foreign_key_test.rb:34-43), so the
  bare names travel Migration#method_missing -> properTableName and the
  WithPrefix / WithSuffix variants actually cover that path.
