+++ packages/activerecord/src/schema-dumper.ts
@@ -1112,9 +1112,9 @@ export abstract class SchemaDumper {
     if (index.where) parts.push(`where: ${JSON.stringify(index.where)}`);
     if (!this._adapter().defaultIndexType(index))
       parts.push(`using: ${JSON.stringify(index.using)}`);
-    if (index.nullsNotDistinct) parts.push("nullsNotDistinct: true");
     if (index.include && index.include.length > 0)
       parts.push(`include: ${JSON.stringify(index.include)}`);
+    if (index.nullsNotDistinct) parts.push("nullsNotDistinct: true");
     // Rails emits `type:` last before `comment:` (schema_dumper.rb#index_parts).
     if (index.type) parts.push(`type: ${JSON.stringify(index.type)}`);
     if (index.comment) parts.push(`comment: ${JSON.stringify(index.comment)}`);
